# new/protocol/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Invity",
    "description": "Volvo API.",
    "version": "0.1"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/project": {
      "get": {
        "summary": "remove last response",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "mentor": {
                  "type": "boolean",
                  "description": "is user a mentor of the project"
                },
                "leader": {
                  "type": "boolean",
                  "description": "is user a leader of the project"
                },
                "member": {
                  "type": "boolean",
                  "description": "is user a member of the project"
                },
                "favourite": {
                  "type": "boolean",
                  "description": "is user's favourite project"
                },
                "recruitment": {
                  "type": "boolean",
                  "description": "is recruitment open"
                },
                "invited": {
                  "type": "boolean",
                  "description": "has been user invited to the project"
                },
                "applied": {
                  "type": "boolean",
                  "description": "has user applied to the project"
                },
                "project": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "depiction": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "image": {
                      "type": "string",
                      "format": "path"
                    }
                  }
                },
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "skill": {
                        "type": "string",
                        "description": "title of skill connected with role"
                      },
                      "members": {
                        "type": "array",
                        "description": "all project members involved in role",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "avatar": {
                              "type": "string",
                              "format": "path"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      },
                      "link": {
                        "type": "string",
                        "format": "url"
                      }
                    }
                  }
                },
                "closure": {
                  "type": "object",
                  "properties": {
                    "closed": {
                      "type": "boolean",
                      "description": "is project closed"
                    },
                    "state": {
                      "type": "boolean",
                      "description": "was closure request sent by the leader"
                    },
                    "explanation": {
                      "type": "string",
                      "description": "the reason why the leader want to close the project"
                    }
                  }
                },
                "response": {
                  "type": "object",
                  "description": "message thet will be shown when user opens a project page; often reffered as 'send message'",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project"
        ]
      }
    },
...
    "/recruitment": {
      "get": {
        "summary": "",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!",
            "schema": {
              "type": "object",
              "properties": {
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "skill": {
                        "type": "string",
                        "format": "skill_code"
                      },
                      "recruitment": {
                        "type": "boolean",
                        "description": "is recruitment open"
                      },
                      "recommended": {
                        "type": "integer",
                        "description": "how meny users have seen this role in theirs recomendations"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      }
                    }
                  }
                },
                "applications": {
                  "type": "array",
                  "description": "all applications for the project",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "surname": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string",
                        "format": "user_id"
                      },
                      "member": {
                        "type": "boolean",
                        "description": "is user a member of the project"
                      },
                      "roles": {
                        "type": "array",
                        "description": "titles of roles they applied",
                        "items": {
                          "type": "string"
                        }
                      },
                      "accepted": {
                        "type": "array",
                        "description": "was user accepted for the according role",
                        "items": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                },
                "skills": {
                  "type": "array",
                  "description": "all available skills",
                  "items": {
                    "type": "object",
                    "properties": {
                      "title": {
                        "type": "string"
                      },
                      "role_title": {
                        "type": "string",
                        "description": "default title of the role (based on the skill title)"
                      },
                      "role_description": {
                        "type": "string",
                        "description": "default description of the role (based on the skill title)"
                      },
                      "icon": {
                        "type": "string",
                        "format": "path"
                      },
                      "hash": {
                        "type": "string",
                        "format": "skill_code"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/change": {
      "get": {
        "summary": "open recruitment for given role if value is true; close recruitment for given role if value if false; if recruitment for any role is open then recruitment for project is open, otherwise it is closed",
        "parameters": [
          {
            "in": "query",
            "name": "hash",
            "type": "string",
            "format": "hash"
          },
          {
            "in": "query",
            "name": "role",
            "type": "string"
          },
          {
            "in": "query",
            "name": "value",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/create_role": {
      "post": {
        "summary": "create role; check if role title is unique, it must be",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/create_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "skill": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/edit_role": {
      "post": {
        "summary": "edit role; check if role title is unique, it must be",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/edit_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newTitle": {
                  "type": "string"
                },
                "oldTitle": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/delete_role": {
      "post": {
        "summary": "delete role; check if nobody is applying for the role, all applications must be dismissed before deleting the role",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/delete_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/accept_role": {
      "post": {
        "summary": "set up a proper notification; add role to user roles; set role in user application as accepted; if user is not a project member, make him a member and send him a welcome message",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/accept_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "role": {
                  "type": "string"
                },
                "response": {
                  "type": "object",
                  "description": "optional welcome message",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/dismiss_role": {
      "post": {
        "summary": "remove role from user application",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/dismiss_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/dismiss_application": {
      "post": {
        "summary": "set up a proper notification; remove whole application and send user a rejection message",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/dismiss_application",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "response": {
                  "type": "object",
                  "description": "reject message",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/involve_member": {
      "post": {
        "summary": "set up a proper notification; add role to member roles; add role to member application; set role in member application as accepted",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/involve_member?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
    "/recruitment/exempt_role": {
      "post": {
        "summary": "set up a proper notification; undo setting member role as accepted",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "[no name] : /recruitment/exempt_role?hash=HASH",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "user_id"
                },
                "role": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success!"
          }
        },
        "tags": [
          "project",
          "project/recruitment"
        ]
      }
    },
...
  },
...
}